final design decision on evaluateLanguagePluginCell, removed evaluate-all-cells.js and migrated the function to actions.js, updated the UserTask for 'run all cells'
