+++ src/actions/actions.js
@@ -165,6 +165,7 @@ function evaluateCodeCell(cell) {
     dispatch(incrementExecutionNumber())
     dispatch(appendToEvalHistory(cell.id, cell.content))
     dispatch(updateUserVariables())
+    // console.log('evaluating code cell', new Date())
   }
 }
 
@@ -272,8 +273,12 @@ function evaluateLanguagePluginCell(cell) {
           dispatch(updateCellProperties(cell.id, { value, evalStatus, rendered }))
           // see the following for asynchronous loading of scripts from strings:
           // https://developer.mozilla.org/en-US/docs/Games/Techniques/Async_scripts
-          var iodideResolvePlugin = resolve // eslint-disable-line
-          var pr = window.eval(xhrObj.responseText) // eslint-disable-line
+
+          // Here, we wrap whatever the return value of the eval into a promise.
+          // If it is simply evaling a code block, then it returns undefined.
+          // But if it returns a Promise, then we can wait for that promise to resolve
+          // before we continue execution.
+          var pr = Promise.resolve(window.eval(xhrObj.responseText)) // eslint-disable-line
 
           pr.then(() => {
             value = `${displayName} plugin ready`
@@ -289,44 +294,8 @@ function evaluateLanguagePluginCell(cell) {
               )
             }
             dispatch(updateCellProperties(cell.id, { value, evalStatus, rendered }))
-            console.log('this should have worked at this point')
             resolve()
           })
-          // const blob = new Blob([xhrObj.responseText])
-          // const script = document.createElement('script')
-
-          // script.id = `plugin-script-${cell.id}`
-
-          // const urlObj = URL.createObjectURL(blob);
-          // script.onload = () => {
-          //   URL.revokeObjectURL(urlObj)
-          //   // NOTE: it is possible to get the blob id used in the browser. for debugging
-          //   // purposes we should be able to map the blob id back to the original filename
-          //   // with just a find/replace in the error output string. this gets the blob id:
-          //   console.log(document.getElementById(`plugin-script-${cell.id}`).src)
-          //   value = `${displayName} plugin ready`
-          //   evalStatus = 'success'
-          //   dispatch(addLanguage(pluginData))
-          //   // FIXME: adding the keybinding move to a reducer ideally, but since it mutates
-          //   // a part of global state in a snowflake sideffect-ish way, and since it
-          //   // needs `dispatch` we'll do it here.
-          //   if (keybinding.length === 1 && (typeof keybinding === 'string')) {
-          //     addLanguageKeybinding(
-          //       [keybinding],
-          //       () => dispatch(changeCellType('code', languageId)),
-          //     )
-          //   }
-          //   dispatch(updateCellProperties(cell.id, { value, evalStatus, rendered }))
-          //   resolve()
-          // }
-          // script.onerror = () => {
-          //   URL.revokeObjectURL(urlObj)
-          //   value = `${displayName} plugin error; script could not be parsed`
-          //   evalStatus = 'error'
-          //   dispatch(updateCellProperties(cell.id, { value, evalStatus, rendered }))
-          // }
-          // script.src = urlObj
-          // document.body.appendChild(script);
         })
 
         xhrObj.addEventListener('error', () => {
@@ -346,7 +315,7 @@ function evaluateLanguagePluginCell(cell) {
 
 export function evaluateCell(cellId) {
   return (dispatch, getState) => {
-    let evaluation// = Promise.resolve()
+    let evaluation
     let cell
     if (cellId === undefined) {
       cell = getSelectedCell(getState())
@@ -370,7 +339,28 @@ export function evaluateCell(cellId) {
     } else {
       cell.rendered = false
     }
-    return evaluation // now it returns a promise. let's see if this works.
+    return evaluation
+  }
+}
+
+export function evaluateAllCells(cells) {
+  return (dispatch) => {
+    let p = Promise.resolve()
+    cells.forEach((cell) => {
+      if (cell.cellType === 'css' && !cell.skipInRunAll) {
+        p = p.then(() => dispatch(evaluateCell(cell.id)))
+      }
+    })
+    cells.forEach((cell) => {
+      if (cell.cellType === 'markdown' && !cell.skipInRunAll) {
+        p = p.then(() => dispatch(evaluateCell(cell.id)))
+      }
+    })
+    cells.forEach((cell) => {
+      if (cell.cellType !== 'markdown' && cell.cellType !== 'css' && !cell.skipInRunAll) {
+        p = p.then(() => dispatch(evaluateCell(cell.id)))
+      }
+    })
   }
 }
